requirements.txt
africastalking==1.2.4
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
django-render-block==0.8.1
django-templated-email==3.0.0
djangorestframework==3.12.4
drf-registration==0.1.2
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
psycopg2==2.9.1
pyparsing==2.4.7
python-dotenv==0.19.1
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
schema==0.7.4
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
